feat:add product entity and controller

+++ src/main/java/webshop/database/repository/ReviewRepository.java
@@ -3,6 +3,6 @@ package webshop.database.repository;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import webshop.database.entity.ReviewDAO;
 
-public interface ReviewRepository extends MongoRepository<ReviewDAO,Integer> {
+public interface ReviewRepository extends MongoRepository<ReviewDAO,Long> {
 
 }
