+++ src/main/java/webshop/database/repository/SupplierRepository.java
@@ -3,6 +3,6 @@ package webshop.database.repository;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import webshop.database.entity.SupplierDAO;
 
-public interface SupplierRepository extends MongoRepository<SupplierDAO,Integer> {
+public interface SupplierRepository extends MongoRepository<SupplierDAO,Long> {
 
 }
